+++ Assets/OxGFrame/CoreFrame/Scripts/Runtime/Core/Implement/FrameManager.cs
@@ -458,26 +458,26 @@ namespace OxGFrame.CoreFrame
         /// 單個關閉
         /// </summary>
         /// <param name="assetName"></param>
-        /// <param name="disableDoSub"></param>
+        /// <param name="disablePreClose"></param>
         /// <param name="forceDestroy"></param>
-        public virtual void Close(string assetName, bool disableDoSub = false, bool forceDestroy = false) { }
+        public virtual void Close(string assetName, bool disablePreClose = false, bool forceDestroy = false) { }
 
         /// <summary>
         /// 全部關閉
         /// </summary>
-        /// <param name="disableDoSub"></param>
+        /// <param name="disablePreClose"></param>
         /// <param name="forceDestroy"></param>
         /// <param name="withoutAssetNames"></param>
-        public virtual void CloseAll(bool disableDoSub = false, bool forceDestroy = false, params string[] withoutAssetNames) { }
+        public virtual void CloseAll(bool disablePreClose = false, bool forceDestroy = false, params string[] withoutAssetNames) { }
 
         /// <summary>
         /// 透過 id 群組進行全部關閉
         /// </summary>
         /// <param name="groupId"></param>
-        /// <param name="disableDoSub"></param>
+        /// <param name="disablePreClose"></param>
         /// <param name="forceDestroy"></param>
         /// <param name="withoutAssetNames"></param>
-        public virtual void CloseAll(int groupId, bool disableDoSub = false, bool forceDestroy = false, params string[] withoutAssetNames) { }
+        public virtual void CloseAll(int groupId, bool disablePreClose = false, bool forceDestroy = false, params string[] withoutAssetNames) { }
         #endregion
 
         #region Reveal