+++ Assets/OxGFrame/CoreFrame/Scripts/Runtime/Core/SRFrame/SRManager.cs
@@ -134,10 +134,10 @@ namespace OxGFrame.CoreFrame.SRFrame
         /// 將 Close 方法封裝 (由接口 Close 與 CloseAll 統一調用)
         /// </summary>
         /// <param name="assetName"></param>
-        /// <param name="disableDoSub"></param>
+        /// <param name="disablePreClose"></param>
         /// <param name="forceDestroy"></param>
         /// <param name="doAll"></param>
-        private void _Close(string assetName, bool disableDoSub, bool forceDestroy, bool doAll)
+        private void _Close(string assetName, bool disablePreClose, bool forceDestroy, bool doAll)
         {
             if (string.IsNullOrEmpty(assetName) || !this.HasStackInAllCache(assetName)) return;
 
@@ -147,7 +147,7 @@ namespace OxGFrame.CoreFrame.SRFrame
                 foreach (var srBase in stack.cache.ToArray())
                 {
                     srBase.SetHidden(false);
-                    this.ExitAndHide(srBase, disableDoSub);
+                    this.ExitAndHide(srBase, disablePreClose);
 
                     if (forceDestroy) this.Destroy(srBase, assetName);
                     else if (srBase.allowInstantiate) this.Destroy(srBase, assetName);
@@ -160,7 +160,7 @@ namespace OxGFrame.CoreFrame.SRFrame
                 if (srBase == null) return;
 
                 srBase.SetHidden(false);
-                this.ExitAndHide(srBase, disableDoSub);
+                this.ExitAndHide(srBase, disablePreClose);
 
                 if (forceDestroy) this.Destroy(srBase, assetName);
                 else if (srBase.allowInstantiate) this.Destroy(srBase, assetName);
@@ -170,14 +170,14 @@ namespace OxGFrame.CoreFrame.SRFrame
             Debug.Log(string.Format("Close SR: 【{0}】", assetName));
         }
 
-        public override void Close(string assetName, bool disableDoSub = false, bool forceDestroy = false)
+        public override void Close(string assetName, bool disablePreClose = false, bool forceDestroy = false)
         {
             // 如果沒有強制 Destroy + 不是顯示狀態則直接 return
             if (!forceDestroy && !this.CheckIsShowing(assetName)) return;
-            this._Close(assetName, disableDoSub, forceDestroy, false);
+            this._Close(assetName, disablePreClose, forceDestroy, false);
         }
 
-        public override void CloseAll(bool disableDoSub = false, bool forceDestroy = false, params string[] withoutAssetNames)
+        public override void CloseAll(bool disablePreClose = false, bool forceDestroy = false, params string[] withoutAssetNames)
         {
             if (this._dictAllCache.Count == 0) return;
 
@@ -209,11 +209,11 @@ namespace OxGFrame.CoreFrame.SRFrame
                 // 如有啟用 CloseAll 需跳過開關, 則不列入關閉執行
                 if (srBase.srSetting.whenCloseAllToSkip) continue;
 
-                this._Close(assetName, disableDoSub, forceDestroy, true);
+                this._Close(assetName, disablePreClose, forceDestroy, true);
             }
         }
 
-        public override void CloseAll(int groupId, bool disableDoSub = false, bool forceDestroy = false, params string[] withoutAssetNames)
+        public override void CloseAll(int groupId, bool disablePreClose = false, bool forceDestroy = false, params string[] withoutAssetNames)
         {
             if (this._dictAllCache.Count == 0) return;
 
@@ -247,7 +247,7 @@ namespace OxGFrame.CoreFrame.SRFrame
                 // 如有啟用 CloseAll 需跳過開關, 則不列入關閉執行
                 if (srBase.srSetting.whenCloseAllToSkip) continue;
 
-                this._Close(assetName, disableDoSub, forceDestroy, true);
+                this._Close(assetName, disablePreClose, forceDestroy, true);
             }
         }
         #endregion
@@ -427,9 +427,9 @@ namespace OxGFrame.CoreFrame.SRFrame
             srBase.Display(obj);
         }
 
-        protected void ExitAndHide(SRBase srBase, bool disableDoSub = false)
+        protected void ExitAndHide(SRBase srBase, bool disablePreClose = false)
         {
-            srBase.Hide(disableDoSub);
+            srBase.Hide(disablePreClose);
         }
         #endregion
     }