+++ Assets/OxGFrame/Samples~/APICenterDemo/Scripts/APICenterExample.cs
@@ -1,7 +1,34 @@
 ﻿using OxGFrame.APICenter;
 
-public class APICenterExample : APICenter<APICenterExample>
+public class APICenterExample : APICenterBase<APICenterExample>
 {
+    #region Default API
+    public static void Add<T>() where T : APIBase, new()
+    {
+        GetInstance().Register<T>();
+    }
+
+    public static void Add<T>(int apiId) where T : APIBase, new()
+    {
+        GetInstance().Register<T>(apiId);
+    }
+
+    public static void Add(int apiId, APIBase apiBase)
+    {
+        GetInstance().Register(apiId, apiBase);
+    }
+
+    public static T Find<T>() where T : APIBase
+    {
+        return GetInstance().GetAPI<T>();
+    }
+
+    public static T Find<T>(int apiId) where T : APIBase
+    {
+        return GetInstance().GetAPI<T>(apiId);
+    }
+    #endregion
+
     public APICenterExample()
     {
         // easy