+++ Assets/OxGFrame/Samples~/SRFrameDemo/Scripts/DemoRS.cs
@@ -11,14 +11,14 @@ public class DemoRS : SRBase
 
     }
 
-    protected override async UniTask OpenSub()
+    protected override async UniTask OnPreShow()
     {
         /**
         * Open Sub With Async
         */
     }
 
-    protected override void CloseSub()
+    protected override void OnPreClose()
     {
         /**
         * Close Sub