+++ Assets/OxGFrame/Samples~/SRFrameDemo/Scripts/DemoSC.cs
@@ -10,14 +10,14 @@ public class DemoSC : SRBase
     {
     }
 
-    protected override async UniTask OpenSub()
+    protected override async UniTask OnPreShow()
     {
         /**
         * Open Sub With Async
         */
     }
 
-    protected override void CloseSub()
+    protected override void OnPreClose()
     {
         /**
         * Close Sub