+++ Assets/OxGFrame/Samples~/UIFrameDemo/Scripts/Demo2UI.cs
@@ -13,14 +13,14 @@ public class Demo2UI : UIBase
     {
     }
 
-    protected override async UniTask OpenSub()
+    protected override async UniTask OnPreShow()
     {
         /**
         * Open Sub With Async
         */
     }
 
-    protected override void CloseSub()
+    protected override void OnPreClose()
     {
         /**
         * Close Sub