+++ Assets/OxGFrame/Samples~/UIFrameDemo/Scripts/Demo3UI.cs
@@ -11,14 +11,14 @@ public class Demo3UI : UIBase
     {
     }
 
-    protected override async UniTask OpenSub()
+    protected override async UniTask OnPreShow()
     {
         /**
         * Open Sub With Async
         */
     }
 
-    protected override void CloseSub()
+    protected override void OnPreClose()
     {
         /**
         * Close Sub