+++ Assets/OxGFrame/Samples~/UIFrameDemo/Scripts/DemoLoadingUI.cs
@@ -7,14 +7,14 @@ public class DemoLoadingUI : UIBase
     {
     }
 
-    protected override async UniTask OpenSub()
+    protected override async UniTask OnPreShow()
     {
         /**
         * Open Sub With Async
         */
     }
 
-    protected override void CloseSub()
+    protected override void OnPreClose()
     {
         /**
         * Close Sub